Refactor TestSupport
